% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_clipboard.R
\name{cb}
\alias{cb}
\title{Copy to clipboard}
\usage{
cb(df, sep = ",", dec = ".", max.size = (200 * 1000))
}
\arguments{
\item{df}{The dataframe that needs to be copied}

\item{sep}{The used character as seperator, (DEFAULT = ",")}

\item{dec}{The used character as decimal, (DEFAULT = ".")}

\item{max.size}{Maximum size of dataframe, (DEFAULT=(200*1000))}
}
\description{
Copy a dataframe to the clipboard
}
