% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colStats.R
\name{colStats}
\alias{colStats}
\title{Statistics per column}
\usage{
colStats(data, func = max)
}
\arguments{
\item{data}{The dataset that needs to be converted}

\item{func}{The function that you want to use}
}
\description{
Get the maximum/minimum/mean/.... from each column in a dataset.
}
