% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_numeric.R
\name{dataframe_numeric}
\alias{dataframe_numeric}
\title{Make a complete dataframe numeric}
\usage{
dataframe_numeric(dataframe)
}
\arguments{
\item{dataframe}{The dataframe that needs to be converted}
}
\description{
Make a complete dataframe numeric. NOTE: this can potentially lead to a data loss and should therefore be handled with care.
}
